Issue #2: Total refactor of project structure
