+++ src/db/state.rs
@@ -1,4 +1,4 @@
-use db::DbState;
+use db_core::db::DbState;
 
 use std::path::Path;
