+++ src/db_core/type_registry.rs
