+++ src/simple_build.rs
@@ -1,0 +1,3 @@
+fn main() {
+    println!("I'm a simple ComposeDB server, but I don't do anything yet.");
+}
